✨ (Inventory Management System): add functionality to check for linked products before deleting categories and suppliers ♻️ (Product, Category, Supplier Controllers): refactor code to include haveLinkedProduct method in controllers and services 📝 (TransactionsPanel): update transaction handling to remove price field as it is no longer needed 💡 (Product): add transaction management methods to handle inventory transactions associated with products

+++ src/main/java/me/yushi/inventorymanagementsystem/view/CategoriesPanel.java
@@ -85,6 +85,10 @@ public class CategoriesPanel extends JPanel {
         }
 
         String categoryID = tableModel.getValueAt(selectedRow, 0).toString();
+        if (categoryController.haveLinkedProduct(categoryID)) {
+            JOptionPane.showMessageDialog(this, "Cannot delete category as it is linked to existing products.", ERROR_MESSAGE, JOptionPane.ERROR_MESSAGE);
+            return;
+        }
         int result = JOptionPane.showConfirmDialog(this, DELETE_CONFIRMATION_MESSAGE, DELETE_CATEGORY_TITLE, JOptionPane.YES_NO_OPTION);
         if (result == JOptionPane.YES_OPTION) {
             categoryController.deleteCategory(categoryID);
